refactor: move ThreadPool from common/ to impl/context/
thread_pool.cpp had a circular dependency with llrt: it lived in common/
but called into MetalContext (which is in impl/) and tt_cluster (in llrt/),
while llrt itself links common. This forced two hacks in common/CMakeLists.txt:

  - A bare include_directories() for tt_metal/llrt/ labelled "Hack - circular dep"
  - add_dependencies(common all_generated_files) to paper over the hidden dep
    on llrt-generated headers (hal/generated/dev_msgs.hpp)

impl/ already legitimately links both llrt (PRIVATE) and common (PUBLIC), so
thread_pool belongs there. Moving it:
  - Removes both hacks from common/CMakeLists.txt
  - Correctly expresses the dependency in the CMake graph
  - Keeps the ThreadPool interface and all callers unchanged

Updated include paths in:
  - tt_metal/impl/profiler/profiler{.cpp,_analysis.hpp}
  - tt_metal/distributed/{mesh_device,mesh_command_queue_base,
    sd/fd/dummy_mesh_command_queue}.{cpp,hpp}
  - tests/tt_metal/distributed/{test,benchmark}_thread_pool.cpp

Builds cleanly (verified locally with clang + ccache).
