+++ tt_metal/common/CMakeLists.txt
@@ -14,7 +14,6 @@ target_include_directories(
     PUBLIC
         ${PROJECT_SOURCE_DIR}
         ${PROJECT_SOURCE_DIR}/tt_metal
-        ${PROJECT_SOURCE_DIR}/tt_metal/llrt # Hack - circular dep, generated dev_msgs.hpp
         ${CMAKE_CURRENT_SOURCE_DIR}
 )
 
@@ -37,9 +36,3 @@ target_link_libraries(
         Boost::asio
         tt-logger::tt-logger
 )
-
-# Ensure generated headers (e.g., tt_metal/llrt/hal/generated/dev_msgs.hpp) exist before compiling this target
-# thread_pool.cpp depends on hal/generated/dev_msgs.hpp
-# Note that this targets dependency on llrt is not captured in cmake
-# This would introduce a circular dependency if we were to add it
-add_dependencies(common all_generated_files)
