+++ tt_metal/distributed/fd_mesh_command_queue.cpp
@@ -39,7 +39,7 @@
 #include <tt_stl/strong_type.hpp>
 #include "dispatch/system_memory_manager.hpp"
 #include "trace/trace_buffer.hpp"
-#include "tt_metal/common/thread_pool.hpp"
+#include "tt_metal/impl/context/thread_pool.hpp"
 #include "tt_metal/common/multi_producer_single_consumer_queue.hpp"
 #include "tt_metal/distributed/mesh_workload_utils.hpp"
 #include "tt_metal/impl/buffers/dispatch.hpp"
